requirements.txt
pika~=1.3.2
python-dotenv~=1.0.1
pyodbc~=5.1.0
python-dateutil~=2.9.0.post0
pymongo~=4.6.2
